+++ plugins/asset-server-plugin/CHANGELOG.md
@@ -1,7 +1,13 @@
 # @shoplyjs/asset-server-plugin
 
+## 2.3.5
+
+### Patch Changes
+
+- remove unused dependency
+
 ## 2.3.4
 
 ### Patch Changes
 
--   f1bb790: Init: cloned from vendure
+- f1bb790: Init: cloned from vendure