+++ plugins/email-plugin/CHANGELOG.md
@@ -1,7 +1,19 @@
 # @shoplyjs/email-plugin
 
+## 2.3.6
+
+### Patch Changes
+
+- test
+
+## 2.3.5
+
+### Patch Changes
+
+- Remove useless dep
+
 ## 2.3.4
 
 ### Patch Changes
 
--   f1bb790: Init: cloned from vendure
+- f1bb790: Init: cloned from vendure