+++ plugins/payments-plugin/CHANGELOG.md
@@ -1,7 +1,13 @@
 # @shoplyjs/payments-plugin
 
+## 2.3.5
+
+### Patch Changes
+
+- init
+
 ## 2.3.4
 
 ### Patch Changes
 
--   f1bb790: Init: cloned from vendure
+- f1bb790: Init: cloned from vendure